pyproject.toml
[project]
name = "cicd"
version = "0.1.0"
description = ""
authors = [
    {name = "Maria Kalikaki",email = "122359543+mariakalykaki@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "cicd", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
